fixed float bug for real this time

+++ harl/envs/isaaclab/Isaac_lab_env.py
@@ -35,7 +35,7 @@ class IsaacLabEnv:
         self.action_space = self.env.action_space
 
     def step(self, actions):
-        actions = torch.tensor(actions.transpose(1, 0, 2), dtype=torch.float64)
+        actions = torch.tensor(actions.transpose(1, 0, 2), dtype=torch.float32)
         obs_all, state_all, reward_all, done_all, info_all, _ = self.env.step(actions)
         return (
             _t2n(obs_all),
